refactor _parsing to other modules
